requirements.txt
numpy
streamlit
pillow
tensorflow==1.15.2
pandas
opencv-python
